# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Q-Wash API
  description: >
    Car-washing queue API: phone + SMS-OTP auth, washing point / service
    CRUD, capacity-aware booking availability, a live queue board, and
    booking history. See `docs/API.md` in the repo for the full prose
    contract and `docs/DATA_MODEL.md` for the schema and algorithms this
    spec is generated alongside.
  version: "1.0.0"
servers:
  - url: /
    description: Same-origin (this server also hosts the docs page)
...
security: []
paths:
...
  /api/v1/auth/otp/request:
    post:
      tags: [Auth]
      operationId: requestOtp
      summary: Request an OTP code
      description: >
        Public. Generates a 6-digit code, bcrypt-hashes it for storage, and
        sends it via the configured `sms.Sender` (a stdout stub in dev — the
        code never appears in the response). Rate-limited: 429
        `otp_cooldown` if an unconsumed code for that phone was issued
        within the last `OTP_COOLDOWN` (default 60s).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                phone_number: { type: string, example: "+15551234567", description: E.164 format }
              required: [phone_number]
      responses:
        "202":
          description: Code generated and "sent"
          content:
            application/json:
              schema:
                type: object
                properties: { status: { type: string, example: otp_sent } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
...
  /api/v1/me/notifications:
    get:
      tags: [Notifications]
      operationId: getMyNotifications
      summary: The caller's own notifications
      description: Paginated, most recently created first.
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/pageSize"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PaginatedNotifications" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/v1/notifications:
    post:
      tags: [Notifications]
      operationId: createNotification
      summary: Send (or schedule) a notification
      description: >
        Requires role `staff` or `admin`. If `queue_id` is given, staff may
        only reference a booking at their own washing point — 404
        `queue_not_found` otherwise. If `send_at` is due (omitted, or
        now/past), sends immediately through the same `sms.Sender` used for
        OTP and persists the outcome (`status: sent`/`failed`). A future
        `send_at` just creates the row as `pending` — there is no
        background worker yet to pick it up later. The call always
        succeeds once the row is created, regardless of delivery outcome.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/NotificationCreate" }
      responses:
        "201":
          description: Created (and possibly sent)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Notification" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /api/v1/connection-requests:
    get:
      tags: [Connection requests]
      operationId: listConnectionRequests
      summary: List connection requests
      description: Requires role admin. Newest first.
      security: [{ bearerAuth: [] }]
      parameters:
        - name: status
          in: query
          schema: { type: string, enum: [new, approved, rejected] }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ConnectionRequestList" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [Connection requests]
      operationId: createConnectionRequest
      summary: Submit a connection request
      description: >
        Requires role admin — admin-created for now, no public self-service
        "apply" form yet.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ConnectionRequestCreate" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ConnectionRequest" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/v1/connection-requests/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [Connection requests]
      operationId: getConnectionRequest
      summary: Get a connection request
      description: Requires role admin.
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ConnectionRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [Connection requests]
      operationId: reviewConnectionRequest
      summary: Approve or reject a connection request
      description: >
        Requires role admin. The only write on an existing request — no
        editing the submitted fields. 409
        `connection_request_already_reviewed` if not still `new`.
        Approving creates an Owner (reusing one whose contact_phone matches,
        if any) and a WashingPoint with status=pending_review and
        placeholder latitude/longitude of 0,0 (the request doesn't collect
        coordinates) — an admin must set them via PATCH
        /washing-points/{id} before the point can go live.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ConnectionRequestReview" }
      responses:
        "200":
          description: Reviewed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ConnectionRequest" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Access token from OTP verify/refresh, e.g. `Authorization: Bearer <token>`"
  parameters:
...
    page:
      name: page
      in: query
      description: 1-indexed. Invalid/out-of-range values clamp to the default rather than erroring.
      schema: { type: integer, minimum: 1, default: 1 }
    pageSize:
      name: page_size
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Missing, malformed, or invalid/expired credentials
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: Authenticated but not permitted
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Not found (or hidden as not-found to avoid leaking existence)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Conflict:
      description: State conflict (e.g. slot no longer available, invalid transition)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    TooManyRequests:
      description: Rate-limited
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string, example: invalid_name }
            message: { type: string, example: "name is required (max 255 chars)" }
          required: [code, message]
      required: [error]
...
    Notification:
      type: object
      properties:
        id: { type: string, format: uuid }
        user_id: { type: string, format: uuid }
        queue_id: { type: string, format: uuid, nullable: true }
        status: { type: string, enum: [pending, sent, failed] }
        channel: { type: string, enum: [sms] }
        text: { type: string }
        send_at: { type: string, format: date-time }
        sent_at: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time }
      required: [id, user_id, status, channel, text, send_at, created_at]
    PaginatedNotifications:
      type: object
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/Notification" } }
        page: { type: integer }
        page_size: { type: integer }
        total: { type: integer }
      required: [items, page, page_size, total]
    NotificationCreate:
      type: object
      properties:
        user_id: { type: string, format: uuid }
        queue_id: { type: string, format: uuid }
        text: { type: string, minLength: 1, maxLength: 2000 }
        send_at: { type: string, format: date-time, description: Defaults to now if omitted }
      required: [user_id, text]
...
    ConnectionRequest:
      type: object
      properties:
        id: { type: string, format: uuid }
        business_name: { type: string }
        contact_name: { type: string }
        contact_phone: { type: string }
        address: { type: string }
        boxes_count: { type: integer, minimum: 1 }
        note: { type: string, nullable: true }
        status: { type: string, enum: [new, approved, rejected] }
        reviewed_by: { type: string, format: uuid, nullable: true }
        reviewed_at: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time }
      required: [id, business_name, contact_name, contact_phone, address, boxes_count, status, created_at]
    ConnectionRequestList:
      type: object
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/ConnectionRequest" } }
      required: [items]
    ConnectionRequestCreate:
      type: object
      properties:
        business_name: { type: string, maxLength: 255 }
        contact_name: { type: string, maxLength: 255 }
        contact_phone: { type: string, maxLength: 32 }
        address: { type: string, maxLength: 500 }
        boxes_count: { type: integer, minimum: 1 }
        note: { type: string, maxLength: 2000 }
      required: [business_name, contact_name, contact_phone, address, boxes_count]
    ConnectionRequestReview:
      type: object
      properties:
        status: { type: string, enum: [approved, rejected] }
      required: [status]
...